requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pyaudio==0.2.13
Pillow==10.0.0
mss==9.0.1
